sidebar-tutorials-not-selected(wi-132611): hack-y fix, but it works
The root cause of the problem is that uri_prefix is "/tutorials" for
*every* tutorial because each tutorial page is served implicitly
from the *directory*, instead of explicitly by their full path. This
is specified in config.json ("/tutorials" is served in "content", which
uses the section_router, and uses markdown_middleware.directory).

Workaround note from routes/section_router.js copied here for reference:

// WORKAROUND NOTE
// Tutorials are rendered implicitly by their directory, instead of
// explicitly by their full path (like the other content types do)
// The uri prefix comes from config.json and is the **dir name**.
// The request path comes from the request itself, which is the
//   rest of the URL excluding the dir name, and may contain a
//   trailing slash.
// req.uriprefix = '/tutorials' || '/maintenance'
// req.path = '/jazzeditor' || '/jazzeditor/' || /jazzeditorjava' || ..etc..